requirements.txt
Django==5.2.7
django-allauth==65.1.0
dj-database-url==2.2.0
gunicorn==23.0.0
whitenoise==6.8.2
Pillow==11.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
